requirements.txt
requests==2.28.2
xmltodict==0.13.0
pyowm==3.3.0
python-telegram-bot==20.2